@@ -262,7 +261,7 @@
 		"vuex": "4.0.0-beta.4",
 		"vuex-persistedstate": "3.0.1",
 		"web-push": "3.4.4",
-		"webpack": "5.0.0-beta.22",
+		"webpack": "git+https://github.com/webpack/webpack.git#c1237eae912817c7546e8c54489f7adb60bfbe38",
 		"webpack-cli": "3.3.12",
 		"websocket": "1.0.31",
 		"ws": "7.3.1",
