@@ -9,7 +9,7 @@
 	</transition>
 
 	<transition name="nav">
-		<nav class="nav" v-show="showing">
+		<nav class="nav" :class="{ iconOnly, hidden }" v-show="showing">
 			<div>
 				<button class="item _button account" @click="openAccountMenu" v-if="$store.getters.isSignedIn">
 					<mk-avatar :user="$store.state.i" class="avatar"/><mk-acct class="text" :user="$store.state.i"/>
@@ -62,6 +62,8 @@ export default defineComponent({
 			menuDef: this.$store.getters.nav({
 				search: this.search
 			}),
+			iconOnly: false,
+			hidden: false,
 			faGripVertical, faChevronLeft, faComments, faHashtag, faBroadcastTower, faFireAlt, faEllipsisH, faPencilAlt, faBars, faTimes, faBell, faSearch, faUserCog, faCog, faUser, faHome, faStar, faCircle, faAt, faEnvelope, faListUl, faPlus, faUserClock, faLaugh, faUsers, faTachometerAlt, faExchangeAlt, faGlobe, faChartBar, faCloud, faServer, faProjectDiagram
 		};
 	},
@@ -85,9 +87,35 @@ export default defineComponent({
 		$route(to, from) {
 			this.showing = false;
 		},
+
+		'$store.state.device.sidebarDisplay'() {
+			this.calcViewState();
+		},
+
+		iconOnly() {
+			this.$nextTick(() => {
+				this.$emit('change-view-mode');
+			});
+		},
+
+		hidden() {
+			this.$nextTick(() => {
+				this.$emit('change-view-mode');
+			});
+		}
+	},
+
+	created() {
+		window.addEventListener('resize', this.calcViewState);
+		this.calcViewState();
 	},
 
 	methods: {
+		calcViewState() {
+			this.iconOnly = (window.innerWidth <= 1279) || (this.$store.state.device.sidebarDisplay === 'icon');
+			this.hidden = (window.innerWidth <= 650) || (this.$store.state.device.sidebarDisplay === 'hide');
+		},
+
 		show() {
 			this.showing = true;
 		},
@@ -314,10 +342,8 @@ export default defineComponent({
 
 .mvcprjjd {
 	$ui-font-size: 1em; // TODO: どこかに集約したい
-	$nav-width: 250px; // TODO: どこかに集約したい
-	$nav-icon-only-width: 80px; // TODO: どこかに集約したい
-	$nav-icon-only-threshold: 1279px; // TODO: どこかに集約したい
-	$nav-hide-threshold: 650px; // TODO: どこかに集約したい
+	$nav-width: 250px;
+	$nav-icon-only-width: 80px;
 
 	> .nav-back {
 		z-index: 1001;
@@ -331,19 +357,66 @@ export default defineComponent({
 		width: $nav-width;
 		box-sizing: border-box;
 
-		@media (max-width: $nav-icon-only-threshold) {
+		&.iconOnly {
 			flex: 0 0 $nav-icon-only-width;
 			width: $nav-icon-only-width;
+
+			&:not(.hidden) {
+				> div {
+					width: $nav-icon-only-width;
+
+					> .divider {
+						margin: 8px auto;
+						width: calc(100% - 32px);
+					}
+
+					> .item {
+						padding-left: 0;
+						width: 100%;
+						text-align: center;
+						font-size: $ui-font-size * 1.1;
+						line-height: 3.7rem;
+
+						> [data-icon],
+						> .avatar {
+							margin-right: 0;
+						}
+
+						> i {
+							left: 10px;
+						}
+
+						> .text {
+							display: none;
+						}
+
+						&:first-child {
+							margin-bottom: 8px;
+						}
+
+						&:last-child {
+							margin-top: 8px;
+						}
+					}
+				}
+			}
 		}
 
-		@media (max-width: $nav-hide-threshold) {
+		&.hidden {
 			position: fixed;
 			top: 0;
 			left: 0;
 			z-index: 1001;
+
+			> div {
+				> .index,
+				> .notifications {
+					display: none;
+				}
+			}
 		}
 
-		@media (min-width: $nav-hide-threshold + 1px) {
+		&:not(.hidden) {
 			display: block !important;
 		}
 
@@ -363,25 +436,6 @@ export default defineComponent({
 			> .divider {
 				margin: 16px 0;
 				border-top: solid 1px var(--divider);
-			}
-
-			@media (max-width: $nav-icon-only-threshold) and (min-width: $nav-hide-threshold + 1px) {
-				width: $nav-icon-only-width;
-
-				> .divider {
-					margin: 8px auto;
-					width: calc(100% - 32px);
-				}
-
-				> .item {
-					&:first-child {
-						margin-bottom: 8px;
-					}
-
-					&:last-child {
-						margin-top: 8px;
-					}
-				}
 			}
 
 			> .item {
@@ -451,34 +505,6 @@ export default defineComponent({
 					bottom: 0;
 					margin-top: 16px;
 					border-top: solid 1px var(--divider);
-				}
-
-				@media (max-width: $nav-icon-only-threshold) and (min-width: $nav-hide-threshold + 1px) {
-					padding-left: 0;
-					width: 100%;
-					text-align: center;
-					font-size: $ui-font-size * 1.2;
-					line-height: 3.7rem;
-
-					> [data-icon],
-					> .avatar {
-						margin-right: 0;
-					}
-
-					> i {
-						left: 10px;
-					}
-
-					> .text {
-						display: none;
-					}
-				}
-			}
-
-			@media (max-width: $nav-hide-threshold) {
-				> .index,
-				> .notifications {
-					display: none;
 				}
 			}
 		}
