@@ -55,16 +55,16 @@ export default defineComponent({
 	},
 	computed: {
 		isMe(): boolean {
-			return this.message.userId == this.$store.state.i.id;
+			return this.message.userId === this.$store.state.i.id;
 		},
 		urls(): string[] {
 			if (this.message.text) {
 				const ast = parse(this.message.text);
 				return unique(ast
-					.filter(t => ((t.node.type == 'url' || t.node.type == 'link') && t.node.props.url && !t.node.props.silent))
+					.filter(t => ((t.node.type === 'url' || t.node.type === 'link') && t.node.props.url && !t.node.props.silent))
 					.map(t => t.node.props.url));
 			} else {
-				return null;
+				return [];
 			}
 		}
 	},
