@@ -13,9 +13,10 @@
 </template>
 
 <script lang="ts">
+import { defineComponent } from 'vue';
 import define from './define';
 
-export default define({
+const widget = define({
 	name: 'digitalClock',
 	props: () => ({
 		transparent: {
@@ -32,7 +33,10 @@ export default define({
 			default: true,
 		},
 	})
-}).extend({
+});
+
+export default defineComponent({
+	extends: widget,
 	data() {
 		return {
 			clock: null,
@@ -45,7 +49,7 @@ export default define({
 	},
 	created() {
 		this.tick();
-		this.$watch('props.showMs', () => {
+		this.$watch(() => this.props.showMs, () => {
 			if (this.clock) clearInterval(this.clock);
 			this.clock = setInterval(this.tick, this.props.showMs ? 10 : 1000);
 		}, { immediate: true });
